Halve the cokebottle weight (and halve the corresponding thrust). The coke bottle no longer can fly out of range this way :-)
This fixes ticket:5.
(It doesn't fix ticket:6 though!)

+++ src/model/CokeMentosBottle.h
@@ -156,7 +156,8 @@ private:
 	/// the amount of liquid still in the bottle
 	/// 2.0l at start, reducing fast whilst blowing.
 	/// unit: kg or l (who cares)
-	qreal	theCokeAmount;
+	/// (note: using the MASS property, you can adjust this)
+	float theCokeAmount;
 
 	/// the weight of an empty bottle in [kg]
 	static const double theBottleMass = 0.2;
